+++ src/modmod/modmodule.h
@@ -72,7 +72,7 @@ private:
 	virtual size_t internal_buildTick( AudioFrameBuffer* buf );
 	virtual std::string internal_channelStatus( size_t idx ) const;
 	virtual std::string internal_channelCellString( size_t idx ) const;
-	virtual uint8_t internal_channelCount() const;
+	virtual int internal_channelCount() const;
 	ModSample* sampleAt( size_t idx ) const;
 	bool existsSample( size_t idx ) const;
 	/**
